+++ rust/operator-binary/src/druid_controller.rs
@@ -262,8 +262,10 @@ pub async fn reconcile_druid(druid: Arc<DruidCluster>, ctx: Arc<Ctx>) -> Result<
         .namespace
         .clone()
         .with_context(|| ObjectHasNoNamespaceSnafu {})?;
-    let resolved_product_image: ResolvedProductImage =
-        druid.spec.image.resolve(DOCKER_IMAGE_BASE_NAME);
+    let resolved_product_image: ResolvedProductImage = druid
+        .spec
+        .image
+        .resolve(DOCKER_IMAGE_BASE_NAME, crate::built_info::CARGO_PKG_VERSION);
 
     let zk_confmap = druid.spec.cluster_config.zookeeper_config_map_name.clone();
     let zk_connstr = client
@@ -1141,8 +1143,10 @@ mod test {
         let druid: DruidCluster =
             serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
 
-        let resolved_product_image: ResolvedProductImage =
-            druid.spec.image.resolve(DOCKER_IMAGE_BASE_NAME);
+        let resolved_product_image: ResolvedProductImage = druid
+            .spec
+            .image
+            .resolve(DOCKER_IMAGE_BASE_NAME, crate::built_info::CARGO_PKG_VERSION);
         let role_config = transform_all_roles_to_config(&druid, druid.build_role_properties());
 
         let product_config_manager =